requirements.txt
psutil>=5.9.0
pandas>=1.5.0
scapy>=2.5.0
yara-python>=4.3.0
requests>=2.31.0
watchdog>=4.0.0